+++ src/YaraParser/TransitionBasedSystem/Configuration/BeamElement.java
@@ -6,12 +6,12 @@
 package YaraParser.TransitionBasedSystem.Configuration;
 
 public class BeamElement implements Comparable<BeamElement> {
-    public float score;
+    public double score;
     public int number;
     public int action;
     public int label;
 
-    public BeamElement(float score, int number, int action, int label) {
+    public BeamElement(double score, int number, int action, int label) {
         this.score = score;
         this.number = number;
         this.action = action;
@@ -20,7 +20,7 @@ public class BeamElement implements Comparable<BeamElement> {
 
     @Override
     public int compareTo(BeamElement beamElement) {
-        float diff = score - beamElement.score;
+        double diff = score - beamElement.score;
         if (diff > 0)
             return 2;
         if (diff < 0)
